Working PWM, Tasker with Timer 1, LedDriver is ready to use with keepAlive led
Pin corrected

+++ trunk/ELE_542_Projet/core/LED/led.h
@@ -33,7 +33,7 @@ public:
 	void deactivateLED();
 	
 	LedType  Type;
-	uint32_t NextEventTime;
+	uint64_t NextEventTime;
 	uint8_t  Active;
 	
 };
